+++ packages/remix_carbon/test/components/carbon_tile_test.dart
@@ -2,6 +2,7 @@ import 'package:remix_carbon/remix_carbon.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:remix/remix.dart';
 
 import '../helpers/pump.dart';
 
@@ -51,6 +52,7 @@ void main() {
       ),
     );
 
+    expect(find.byType(RemixDisclosure), findsOneWidget);
     expect(find.text('Hidden details'), findsNothing);
     await tester.tap(find.bySemanticsLabel('Details'));
     await tester.pump();