+++ packages/remix_carbon/test/components/carbon_ui_shell_test.dart
@@ -3,6 +3,7 @@ import 'dart:ui' show SemanticsRole;
 import 'package:remix_carbon/remix_carbon.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:remix/remix.dart';
 
 import '../helpers/pump.dart';
 
@@ -28,6 +29,7 @@ void main() {
       ),
     );
 
+    expect(find.byType(RemixSidebar<int>), findsOneWidget);
     expect(find.text('Workspace'), findsOneWidget);
     expect(
       tester.getSemantics(find.bySemanticsLabel('Primary navigation')).role,